pyproject.toml
[tool.poetry]
name = "mentormingle"
version = "0.1.0"
description = "An AI Agent for interview prep"
authors = ["Aymane Hachcham & Sebti Youssef"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.8.6"
aiosignal = "1.3.1"
async-timeout = "4.0.3"
attrs = "23.1.0"
certifi = "2023.7.22"
charset-normalizer = "3.3.0"
frozenlist = "1.4.0"
idna = "3.4"
load-dotenv = "0.1.0"
multidict = "6.0.4"
openai = "0.28.1"
python-dotenv = "1.0.0"
requests = "2.31.0"
tqdm = "4.66.1"
urllib3 = "2.0.6"
yarl = "1.9.2"
pydantic = "^2.4.2"
toml = "^0.10.2"
pydantic-settings = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
